does not delete while searching

+++ Charge/Controllers/DashboardTableViewController.swift
@@ -123,7 +123,7 @@ extension DashboardTableViewController {
     }
 
     override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
-        if editingStyle == .delete {
+        if editingStyle == .delete  && !self.resultSearchController.isActive{
             CoreService.delete(xCObject: self.items[indexPath.item])
             self.refresh()
         }
